pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Application API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.6.2.post1",
    "certifi>=2024.8.30",
    "click>=8.1.7",
    "colorama>=0.4.6",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "exceptiongroup>=1.2.2",
    "fastapi>=0.115.5",
    "fastapi-cli>=0.0.5",
    "greenlet>=3.1.1",
    "h11>=0.14.0",
    "httpcore>=1.0.7",
    "httptools>=0.6.4",
    "httpx>=0.27.2",
    "idna>=3.10",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.4",
    "markdown-it-py>=3.0.0",
    "markupsafe>=3.0.2",
    "mdurl>=0.1.2",
    "orjson>=3.10.11",
    "pydantic>=2.10.0",
    "pydantic-core>=2.27.2",
    "pydantic-extra-types>=2.10.0",
    "pydantic-settings>=2.6.1",
    "pygments>=2.18.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "shellingham>=1.5.4",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.36",
    "starlette>=0.41.3",
    "typer>=0.13.",
    "typing-extensions>=4.12.2",
    "ujson>=5.10.0",
    "uvicorn>=0.32.1",
    "watchfiles>=0.24.0",
    "websockets>=14.1",
    "cffi>=1.17.1",
    "cryptography>=43.0.3",
    "ecdsa>=0.19.0",
    "pyasn1>=0.6.1",
    "pycparser>=2.22",
    "python-jose>=3.3.0",
    "rsa>=4.9",
    "six>=1.16.0",
    "bcrypt>=4.2.1",
    "passlib>=1.7.4",
    "python-multipart>=0.0.17",
    "alembic>=1.14.0",
    "mako>=1.3.8",
    "redis>=5.2.1",
    "aiosmtplib>=3.0.2",
    "blinker>=1.9.0",
    "fastapi-mail>=1.4.2"
]
